requirements.txt
python-dotenv==1.0.0
aiohttp==3.8.5
tenacity==8.2.2
pydantic==2.0.3
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-mock==3.11.1